+++ Content.Server/_Sunrise/Wires/BaseWireAction.cs
@@ -1,0 +1,15 @@
+namespace Content.Server.Wires;
+using Content.Shared.Doors.Components;
+
+public abstract partial class BaseWireAction : IWireAction
+{
+    public void WireCutSparks(EntityUid uid)
+    {
+        if (!IsPowered(uid))
+            return;
+        if (!EntityManager.TryGetComponent<DoorComponent>(uid, out var door) || !door.WireCutSparks)
+            return;
+        var coords = EntityManager.GetComponent<TransformComponent>(uid).Coordinates;
+        EntityManager.SpawnEntity("EffectMechSparks", coords);
+    }
+}
