+++ Content.Shared/_Sunrise/Doors/Components/DoorComponent.cs
@@ -1,0 +1,7 @@
+namespace Content.Shared.Doors.Components;
+
+public sealed partial class DoorComponent : Component
+{
+    [DataField]
+    public bool WireCutSparks = true;
+}
